+++ src/components/popups/Dropdown.tsx
@@ -5,11 +5,11 @@ import {Popups} from '../../common/interfaces';
 import {bemClasses} from '../../common/bem';
 import {List} from '../lists';
 
-export const Dropdown: React.FC<Popups.IDropdown & IPopupProps> = ({isOpen, items = [], ...props}) => {
+export const Dropdown: React.FC<Popups.IDropdown & IPopupProps> = ({isOpen, zIndex, items = [], ...props}) => {
   const classBase = 'rcy-dropdown';
   const className = bemClasses(classBase, props.mods, props.className);
   return (
-    <div className={className}>
+    <div className={className} style={{zIndex}}>
       <List
         className={`${classBase}__list ${!isOpen ? 'rcy-hidden' : ''}`}
         itemClassName={`${classBase}__item ${props.itemClassName ? props.itemClassName : ''}`}