+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 * Select and dropdown
 ### 1.0.90
 * Popup manager
+### 1.0.93
+* Popup manager fix
+### 1.0.95
+* Popup manager zIndex